speech_analyzer/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
openai-whisper==20231117
ffmpeg-python==0.2.0
python-multipart==0.0.9
boto3==1.34.0
numpy
torch
